# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/ragflow/search:
        post:
            tags:
                - RAGFlowService
            description: SemanticSearch 执行语义检索
            operationId: RAGFlowService_SemanticSearch
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SemanticSearchRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SemanticSearchResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ragflow/stats:
        get:
            tags:
                - RAGFlowService
            description: GetSyncStats 获取同步统计信息
            operationId: RAGFlowService_GetSyncStats
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SyncStats'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ragflow/status:
        get:
            tags:
                - RAGFlowService
            description: GetSyncStatus 获取 RAGFlow 同步服务状态
            operationId: RAGFlowService_GetSyncStatus
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SyncStatus'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ragflow/sync:
        post:
            tags:
                - RAGFlowService
            description: TriggerSync 手动触发一次同步
            operationId: RAGFlowService_TriggerSync
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TriggerSyncRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ragflow/sync-states:
        get:
            tags:
                - RAGFlowService
            description: ListContentSyncStates 列出内容同步状态
            operationId: RAGFlowService_ListContentSyncStates
            parameters:
                - name: statusFilter
                  in: query
                  description: '过滤状态: pending, synced, failed, skipped'
                  schema:
                    type: string
                - name: contentTypeFilter
                  in: query
                  description: '内容类型过滤: memo, attachment'
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  description: 分页大小
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  description: 分页令牌
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListContentSyncStatesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        ContentSyncState:
            type: object
            properties:
                contentType:
                    type: string
                    description: '内容类型: memo, attachment'
                contentUid:
                    type: string
                    description: 内容 UID
                status:
                    type: string
                    description: '同步状态: pending, synced, failed, skipped'
                ragflowDocumentId:
                    type: string
                    description: RAGFlow 文档 ID
                syncedTime:
                    type: string
                    description: 同步时间
                    format: date-time
                errorMessage:
                    type: string
                    description: 错误信息（如果失败）
                retryCount:
                    type: integer
                    description: 重试次数
                    format: int32
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ListContentSyncStatesResponse:
            type: object
            properties:
                syncStates:
                    type: array
                    items:
                        $ref: '#/components/schemas/ContentSyncState'
                    description: 同步状态列表
                nextPageToken:
                    type: string
                    description: 下一页令牌
...
        SearchResult:
            type: object
            properties:
                memoName:
                    type: string
                    description: Memo 资源名称 (memos/{uid})
                contentSnippet:
                    type: string
                    description: Memo 内容片段
                similarityScore:
                    type: number
                    description: 相似度分数 (0.0-1.0)
                    format: float
                title:
                    type: string
                    description: Memo 标题（第一行或前50字符）
                createTime:
                    type: string
                    description: 创建时间
                    format: date-time
        SemanticSearchRequest:
            required:
                - query
            type: object
            properties:
                query:
                    type: string
                    description: 搜索查询
                topK:
                    type: integer
                    description: 返回结果数量限制
                    format: int32
                similarityThreshold:
                    type: number
                    description: 相似度阈值 (0.0-1.0)
                    format: float
        SemanticSearchResponse:
            type: object
            properties:
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/SearchResult'
                    description: 搜索结果列表
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        SyncStats:
            type: object
            properties:
                pendingCount:
                    type: integer
                    description: 待同步数量
                    format: int32
                syncedCount:
                    type: integer
                    description: 已同步数量
                    format: int32
                failedCount:
                    type: integer
                    description: 失败数量
                    format: int32
                skippedCount:
                    type: integer
                    description: 跳过数量
                    format: int32
                totalCount:
                    type: integer
                    description: 总数量
                    format: int32
        SyncStatus:
            type: object
            properties:
                healthy:
                    type: boolean
                    description: RAGFlow 服务是否健康
                circuitOpen:
                    type: boolean
                    description: 熔断器是否打开
                lastCheckTime:
                    type: string
                    description: 最后检查时间
                    format: date-time
                failureCount:
                    type: integer
                    description: 连续失败次数
                    format: int32
                runnerActive:
                    type: boolean
                    description: 同步服务是否运行中
        TriggerSyncRequest:
            type: object
            properties: {}
...